requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
requests==2.27.1
urllib3==1.26.9
mariadb==1.0.11
python-telegram-bot==13.8.1
python-dotenv==0.18.0
